issue 990 move v2 api to v1
