+++ sqle/api/controller/v1/workflow.go
@@ -2,6 +2,7 @@ package v1
 
 import (
 	"database/sql"
+	_err "errors"
 	"fmt"
 	"net/http"
 	"strconv"
@@ -19,6 +20,15 @@ import (
 )
 
 var ErrWorkflowNoAccess = errors.New(errors.DataNotExist, fmt.Errorf("workflow is not exist or you can't access it"))
+
+var ErrForbidMyBatisXMLTask = func(taskId uint) error {
+	return errors.New(errors.DataConflict,
+		fmt.Errorf("the task for audit mybatis xml file is not allow to create workflow. taskId=%v", taskId))
+}
+
+var ErrWorkflowExecuteTimeIncorrect = errors.New(errors.TaskActionInvalid, fmt.Errorf("please go online during instance operation and maintenance time"))
+
+var errTaskHasBeenUsed = errors.New(errors.DataConflict, fmt.Errorf("task has been used in other workflow"))
 
 type GetWorkflowTemplateResV1 struct {
 	controller.BaseRes
@@ -533,43 +543,6 @@ func CheckCurrentUserCanOperateWorkflow(c echo.Context, workflow *model.Workflow
 	return ErrWorkflowNoAccess
 }
 
-type GetWorkflowsReqV1 struct {
-	FilterSubject                     string `json:"filter_subject" query:"filter_subject"`
-	FilterCreateTimeFrom              string `json:"filter_create_time_from" query:"filter_create_time_from"`
-	FilterCreateTimeTo                string `json:"filter_create_time_to" query:"filter_create_time_to"`
-	FilterCreateUserName              string `json:"filter_create_user_name" query:"filter_create_user_name"`
-	FilterCurrentStepType             string `json:"filter_current_step_type" query:"filter_current_step_type" valid:"omitempty,oneof=sql_review sql_execute"`
-	FilterStatus                      string `json:"filter_status" query:"filter_status" valid:"omitempty,oneof=on_process rejected canceled exec_scheduled executing exec_failed finished"`
-	FilterCurrentStepAssigneeUserName string `json:"filter_current_step_assignee_user_name" query:"filter_current_step_assignee_user_name"`
-	FilterTaskInstanceName            string `json:"filter_task_instance_name" query:"filter_task_instance_name"`
-	FilterTaskExecuteStartTimeFrom    string `json:"filter_task_execute_start_time_from" query:"filter_task_execute_start_time_from"`
-	FilterTaskExecuteStartTimeTo      string `json:"filter_task_execute_start_time_to" query:"filter_task_execute_start_time_to"`
-	PageIndex                         uint32 `json:"page_index" query:"page_index" valid:"required"`
-	PageSize                          uint32 `json:"page_size" query:"page_size" valid:"required"`
-}
-
-type GetWorkflowsResV1 struct {
-	controller.BaseRes
-	Data      []*WorkflowDetailResV1 `json:"data"`
-	TotalNums uint64                 `json:"total_nums"`
-}
-
-type WorkflowDetailResV1 struct {
-	Id                      uint       `json:"workflow_id"`
-	Subject                 string     `json:"subject"`
-	Desc                    string     `json:"desc"`
-	TaskPassRate            float64    `json:"task_pass_rate"`
-	TaskScore               int32      `json:"task_score"`
-	TaskInstance            string     `json:"task_instance_name"`
-	TaskInstanceSchema      string     `json:"task_instance_schema"`
-	CreateUser              string     `json:"create_user_name"`
-	CreateTime              *time.Time `json:"create_time"`
-	CurrentStepType         string     `json:"current_step_type,omitempty" enums:"sql_review,sql_execute"`
-	CurrentStepAssigneeUser []string   `json:"current_step_assignee_user_name_list,omitempty"`
-	Status                  string     `json:"status" enums:"on_process,rejected,canceled,exec_scheduled,executing,exec_failed,finished"`
-	ScheduleTime            *time.Time `json:"schedule_time,omitempty"`
-}
-
 func CheckUserCanOperateStep(user *model.User, workflow *model.Workflow, stepId int) error {
 	if workflow.Record.Status != model.WorkflowStatusWaitForAudit && workflow.Record.Status != model.WorkflowStatusWaitForExecution {
 		return fmt.Errorf("workflow status is %s, not allow operate it", workflow.Record.Status)
@@ -847,10 +820,6 @@ func checkCancelWorkflow(id string) (*model.Workflow, error) {
 	return workflow, nil
 }
 
-type UpdateWorkflowReqV1 struct {
-	TaskId string `json:"task_id" form:"task_id" valid:"required"`
-}
-
 func FormatStringToInt(s string) (ret int, err error) {
 	if s == "" {
 		return 0, nil
@@ -873,10 +842,6 @@ func FormatStringToUint64(s string) (ret uint64, err error) {
 		}
 	}
 	return ret, nil
-}
-
-type UpdateWorkflowScheduleV1 struct {
-	ScheduleTime *time.Time `json:"schedule_time"`
 }
 
 // ExecuteOneTaskOnWorkflowV1
@@ -1152,4 +1117,562 @@ func CheckCurrentUserCanViewWorkflow(c echo.Context, workflow *model.Workflow) e
 		return nil
 	}
 	return ErrWorkflowNoAccess
+}
+
+type CreateWorkflowReqV1 struct {
+	Subject string `json:"workflow_subject" form:"workflow_subject" valid:"required,name"`
+	Desc    string `json:"desc" form:"desc"`
+	TaskIds []uint `json:"task_ids" form:"task_ids" valid:"required"`
+}
+
+// CreateWorkflowV1
+// @Summary 创建工单
+// @Description create workflow
+// @Accept json
+// @Produce json
+// @Tags workflow
+// @Id createWorkflowV1
+// @Security ApiKeyAuth
+// @Param instance body v1.CreateWorkflowReqV1 true "create workflow request"
+// @Param project_id path uint true "project id"
+// @Success 200 {object} controller.BaseRes
+// @router /v1/projects/{project_id}/workflows [post]
+func CreateWorkflowV1(c echo.Context) error {
+	req := new(CreateWorkflowReqV1)
+	if err := controller.BindAndValidateReq(c, req); err != nil {
+		return err
+	}
+	s := model.GetStorage()
+
+	_, exist, err := s.GetWorkflowBySubject(req.Subject)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if exist {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataExist, fmt.Errorf("workflow is exist")))
+	}
+
+	taskIds := utils.RemoveDuplicateUint(req.TaskIds)
+	if len(taskIds) > MaximumDataSourceNum {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataConflict, fmt.Errorf("the max task count of a workflow is %v", MaximumDataSourceNum)))
+	}
+	tasks, foundAllTasks, err := s.GetTasksByIds(taskIds)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !foundAllTasks {
+		return controller.JSONBaseErrorReq(c, ErrTaskNoAccess)
+	}
+
+	user, err := controller.GetCurrentUser(c)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	workflowTemplateId := tasks[0].Instance.WorkflowTemplateId
+	for _, task := range tasks {
+		if task.Instance == nil {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataNotExist, fmt.Errorf("instance is not exist. taskId=%v", task.ID)))
+		}
+
+		count, err := s.GetTaskSQLCountByTaskID(task.ID)
+		if err != nil {
+			return controller.JSONBaseErrorReq(c, err)
+		}
+		if count == 0 {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, fmt.Errorf("workflow's execute sql is null. taskId=%v", task.ID)))
+		}
+
+		if task.CreateUserId != user.ID {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataConflict,
+				fmt.Errorf("the task is not created by yourself. taskId=%v", task.ID)))
+		}
+
+		if task.SQLSource == model.TaskSQLSourceFromMyBatisXMLFile {
+			return controller.JSONBaseErrorReq(c, ErrForbidMyBatisXMLTask(task.ID))
+		}
+
+		// all instances must use the same workflow template
+		if task.Instance.WorkflowTemplateId != workflowTemplateId {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataConflict,
+				fmt.Errorf("all instances must use the same workflow template")))
+		}
+	}
+
+	// check user role operations
+	{
+		err = checkCurrentUserCanCreateWorkflow(user, tasks)
+		if err != nil {
+			return controller.JSONBaseErrorReq(c, err)
+		}
+	}
+
+	count, err := s.GetWorkflowRecordCountByTaskIds(taskIds)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if count > 0 {
+		return controller.JSONBaseErrorReq(c, errTaskHasBeenUsed)
+	}
+
+	template, exist, err := s.GetWorkflowTemplateById(workflowTemplateId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataNotExist,
+			fmt.Errorf("the task instance is not bound workflow template")))
+	}
+
+	err = checkWorkflowCanCommit(template, tasks)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	stepTemplates, err := s.GetWorkflowStepsByTemplateId(template.ID)
+	if err != nil {
+		return err
+	}
+	err = s.CreateWorkflow(req.Subject, req.Desc, user, tasks, stepTemplates)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	workflow, exist, err := s.GetLastWorkflow()
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataNotExist, fmt.Errorf("should exist at least one workflow after create workflow")))
+	}
+	go notification.NotifyWorkflow(fmt.Sprintf("%v", workflow.ID), notification.WorkflowNotifyTypeCreate)
+
+	return c.JSON(http.StatusOK, controller.NewBaseReq(nil))
+}
+
+func checkWorkflowCanCommit(template *model.WorkflowTemplate, tasks []*model.Task) error {
+	allowLevel := driver.RuleLevelError
+	if template.AllowSubmitWhenLessAuditLevel != "" {
+		allowLevel = driver.RuleLevel(template.AllowSubmitWhenLessAuditLevel)
+	}
+	for _, task := range tasks {
+		if driver.RuleLevel(task.AuditLevel).More(allowLevel) {
+			return errors.New(errors.DataInvalid,
+				fmt.Errorf("there is an audit result with an error level higher than the allowable submission level(%v), please modify it before submitting. taskId=%v", allowLevel, task.ID))
+		}
+	}
+	return nil
+}
+
+func checkCurrentUserCanCreateWorkflow(user *model.User, tasks []*model.Task) error {
+	if model.IsDefaultAdminUser(user.Name) {
+		return nil
+	}
+
+	instances := make([]*model.Instance, len(tasks))
+	for i, task := range tasks {
+		instances[i] = task.Instance
+	}
+
+	s := model.GetStorage()
+	ok, err := s.CheckUserHasOpToInstances(user, instances, []uint{model.OP_WORKFLOW_SAVE})
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return errors.NewAccessDeniedErr("user has no access to create workflow for instance")
+	}
+
+	return nil
+}
+
+type GetWorkflowsReqV1 struct {
+	FilterSubject                     string `json:"filter_subject" query:"filter_subject"`
+	FilterCreateTimeFrom              string `json:"filter_create_time_from" query:"filter_create_time_from"`
+	FilterCreateTimeTo                string `json:"filter_create_time_to" query:"filter_create_time_to"`
+	FilterCreateUserName              string `json:"filter_create_user_name" query:"filter_create_user_name"`
+	FilterStatus                      string `json:"filter_status" query:"filter_status" valid:"omitempty,oneof=wait_for_audit wait_for_execution rejected canceled executing exec_failed finished"`
+	FilterCurrentStepAssigneeUserName string `json:"filter_current_step_assignee_user_name" query:"filter_current_step_assignee_user_name"`
+	FilterTaskInstanceName            string `json:"filter_task_instance_name" query:"filter_task_instance_name"`
+	FilterTaskExecuteStartTimeFrom    string `json:"filter_task_execute_start_time_from" query:"filter_task_execute_start_time_from"`
+	FilterTaskExecuteStartTimeTo      string `json:"filter_task_execute_start_time_to" query:"filter_task_execute_start_time_to"`
+	PageIndex                         uint32 `json:"page_index" query:"page_index" valid:"required"`
+	PageSize                          uint32 `json:"page_size" query:"page_size" valid:"required"`
+}
+
+type GetWorkflowsResV1 struct {
+	controller.BaseRes
+	Data      []*WorkflowDetailResV1 `json:"data"`
+	TotalNums uint64                 `json:"total_nums"`
+}
+
+type WorkflowDetailResV1 struct {
+	Id                      uint       `json:"workflow_id"`
+	Subject                 string     `json:"subject"`
+	Desc                    string     `json:"desc"`
+	CreateUser              string     `json:"create_user_name"`
+	CreateTime              *time.Time `json:"create_time"`
+	CurrentStepType         string     `json:"current_step_type,omitempty" enums:"sql_review,sql_execute"`
+	CurrentStepAssigneeUser []string   `json:"current_step_assignee_user_name_list,omitempty"`
+	Status                  string     `json:"status" enums:"wait_for_audit,wait_for_execution,rejected,canceled,exec_failed,executing,finished"`
+}
+
+// GetWorkflowsV1
+// @Summary 获取工单列表
+// @Description get workflow list
+// @Tags workflow
+// @Id getWorkflowsV1
+// @Security ApiKeyAuth
+// @Param filter_subject query string false "filter subject"
+// @Param filter_create_time_from query string false "filter create time from"
+// @Param filter_create_time_to query string false "filter create time to"
+// @Param filter_task_execute_start_time_from query string false "filter_task_execute_start_time_from"
+// @Param filter_task_execute_start_time_to query string false "filter_task_execute_start_time_to"
+// @Param filter_create_user_name query string false "filter create user name"
+// @Param filter_status query string false "filter workflow status" Enums(wait_for_audit,wait_for_execution,rejected,executing,canceled,exec_failed,finished)
+// @Param filter_current_step_assignee_user_name query string false "filter current step assignee user name"
+// @Param filter_task_instance_name query string false "filter instance name"
+// @Param page_index query uint32 false "page index"
+// @Param page_size query uint32 false "size of per page"
+// @Param project_id path uint true "project id"
+// @Success 200 {object} v1.GetWorkflowsResV1
+// @router /v1/projects/{project_id}/workflows [get]
+func GetWorkflowsV1(c echo.Context) error {
+	req := new(GetWorkflowsReqV1)
+	if err := controller.BindAndValidateReq(c, req); err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	user, err := controller.GetCurrentUser(c)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	var offset uint32
+	if req.PageIndex > 0 {
+		offset = (req.PageIndex - 1) * req.PageSize
+	}
+
+	data := map[string]interface{}{
+		"filter_subject":                         req.FilterSubject,
+		"filter_create_time_from":                req.FilterCreateTimeFrom,
+		"filter_create_time_to":                  req.FilterCreateTimeTo,
+		"filter_create_user_name":                req.FilterCreateUserName,
+		"filter_task_execute_start_time_from":    req.FilterTaskExecuteStartTimeFrom,
+		"filter_task_execute_start_time_to":      req.FilterTaskExecuteStartTimeTo,
+		"filter_status":                          req.FilterStatus,
+		"filter_current_step_assignee_user_name": req.FilterCurrentStepAssigneeUserName,
+		"filter_task_instance_name":              req.FilterTaskInstanceName,
+		"current_user_id":                        user.ID,
+		"check_user_can_access":                  user.Name != model.DefaultAdminUser,
+		"limit":                                  req.PageSize,
+		"offset":                                 offset,
+	}
+
+	s := model.GetStorage()
+	workflows, count, err := s.GetWorkflowsByReq(data, user)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	workflowsReq := make([]*WorkflowDetailResV1, 0, len(workflows))
+	for _, workflow := range workflows {
+		workflowReq := &WorkflowDetailResV1{
+			Id:                      workflow.Id,
+			Subject:                 workflow.Subject,
+			Desc:                    workflow.Desc,
+			CreateUser:              utils.AddDelTag(workflow.CreateUserDeletedAt, workflow.CreateUser.String),
+			CreateTime:              workflow.CreateTime,
+			CurrentStepType:         workflow.CurrentStepType.String,
+			CurrentStepAssigneeUser: workflow.CurrentStepAssigneeUser,
+			Status:                  workflow.Status,
+		}
+		workflowsReq = append(workflowsReq, workflowReq)
+	}
+
+	return c.JSON(http.StatusOK, GetWorkflowsResV1{
+		BaseRes:   controller.NewBaseReq(nil),
+		Data:      workflowsReq,
+		TotalNums: count,
+	})
+}
+
+type UpdateWorkflowReqV1 struct {
+	TaskIds []uint `json:"task_ids" form:"task_ids" valid:"required"`
+}
+
+// UpdateWorkflowV1
+// @Summary 更新工单（驳回后才可更新）
+// @Description update workflow when it is rejected to creator.
+// @Tags workflow
+// @Accept json
+// @Produce json
+// @Id updateWorkflowV1
+// @Security ApiKeyAuth
+// @Param workflow_id path string true "workflow id"
+// @Param instance body v1.UpdateWorkflowReqV1 true "update workflow request"
+// @Success 200 {object} controller.BaseRes
+// @router /v1/projects/workflows/{workflow_id}/ [patch]
+func UpdateWorkflowV1(c echo.Context) error {
+	req := new(UpdateWorkflowReqV1)
+	if err := controller.BindAndValidateReq(c, req); err != nil {
+		return err
+	}
+	workflowIdStr := c.Param("workflow_id")
+	workflowId, err := FormatStringToInt(workflowIdStr)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	err = CheckCurrentUserCanOperateWorkflow(c, &model.Workflow{
+		Model: model.Model{ID: uint(workflowId)},
+	}, []uint{})
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	s := model.GetStorage()
+	tasks, _, err := s.GetTasksByIds(req.TaskIds)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if len(tasks) <= 0 {
+		return controller.JSONBaseErrorReq(c, ErrTaskNoAccess)
+	}
+
+	user, err := controller.GetCurrentUser(c)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	taskIds := make([]uint, len(tasks))
+	for i, task := range tasks {
+		taskIds[i] = task.ID
+
+		count, err := s.GetTaskSQLCountByTaskID(task.ID)
+		if err != nil {
+			return controller.JSONBaseErrorReq(c, err)
+		}
+		if count == 0 {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, fmt.Errorf("task's execute sql is null. taskId=%v", task.ID)))
+		}
+
+		err = CheckCurrentUserCanViewTask(c, task)
+		if err != nil {
+			return controller.JSONBaseErrorReq(c, err)
+		}
+
+		if task.Instance == nil {
+			return controller.JSONBaseErrorReq(c, ErrInstanceNotExist)
+		}
+
+		if user.ID != task.CreateUserId {
+			return controller.JSONBaseErrorReq(c, errors.New(errors.DataConflict,
+				fmt.Errorf("the task is not created by yourself. taskId=%v", task.ID)))
+		}
+
+		if task.SQLSource == model.TaskSQLSourceFromMyBatisXMLFile {
+			return controller.JSONBaseErrorReq(c, ErrForbidMyBatisXMLTask(task.ID))
+		}
+	}
+
+	count, err := s.GetWorkflowRecordCountByTaskIds(taskIds)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if count > 0 {
+		return controller.JSONBaseErrorReq(c, errTaskHasBeenUsed)
+	}
+
+	workflow, exist, err := s.GetWorkflowDetailById(workflowIdStr)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, ErrWorkflowNoAccess)
+	}
+
+	if workflow.Record.Status != model.WorkflowStatusReject {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid,
+			fmt.Errorf("workflow status is %s, not allow operate it", workflow.Record.Status)))
+	}
+
+	if user.ID != workflow.CreateUserId {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataNotExist,
+			fmt.Errorf("you are not allow to operate the workflow")))
+	}
+
+	template, exist, err := s.GetWorkflowTemplateById(tasks[0].Instance.WorkflowTemplateId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataConflict,
+			fmt.Errorf("failed to find the corresponding workflow template based on the task id")))
+	}
+
+	err = checkWorkflowCanCommit(template, tasks)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	stepTemplates, err := s.GetWorkflowStepsByTemplateId(template.ID)
+	if err != nil {
+		return err
+	}
+
+	err = s.UpdateWorkflowRecord(workflow, tasks, stepTemplates)
+	if err != nil {
+		return c.JSON(http.StatusOK, controller.NewBaseReq(err))
+	}
+	go notification.NotifyWorkflow(workflowIdStr, notification.WorkflowNotifyTypeCreate)
+
+	return c.JSON(http.StatusOK, controller.NewBaseReq(nil))
+}
+
+type UpdateWorkflowScheduleReqV1 struct {
+	ScheduleTime *time.Time `json:"schedule_time"`
+}
+
+// UpdateWorkflowScheduleV1
+// @Summary 设置工单数据源定时上线时间（设置为空则代表取消定时时间，需要SQL审核流程都通过后才可以设置）
+// @Description update workflow schedule.
+// @Tags workflow
+// @Accept json
+// @Produce json
+// @Id updateWorkflowScheduleV1
+// @Security ApiKeyAuth
+// @Param workflow_id path string true "workflow id"
+// @Param task_id path string true "task id"
+// @Param instance body v1.UpdateWorkflowScheduleReqV1 true "update workflow schedule request"
+// @Success 200 {object} controller.BaseRes
+// @router /v1/projects/workflows/{workflow_id}/tasks/{task_id}/schedule [put]
+func UpdateWorkflowScheduleV1(c echo.Context) error {
+	workflowId := c.Param("workflow_id")
+	workflowIdInt, err := FormatStringToInt(workflowId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	taskId := c.Param("task_id")
+	taskIdUint, err := FormatStringToUint64(taskId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	req := new(UpdateWorkflowScheduleReqV1)
+	if err := controller.BindAndValidateReq(c, req); err != nil {
+		return err
+	}
+
+	err = CheckCurrentUserCanOperateWorkflow(c, &model.Workflow{
+		Model: model.Model{ID: uint(workflowIdInt)},
+	}, []uint{})
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	user, err := controller.GetCurrentUser(c)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	s := model.GetStorage()
+	workflow, exist, err := s.GetWorkflowDetailById(workflowId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, ErrWorkflowNoAccess)
+	}
+	currentStep := workflow.CurrentStep()
+	if currentStep == nil {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, _err.New("workflow current step not found")))
+	}
+
+	if workflow.Record.Status != model.WorkflowStatusWaitForExecution {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid,
+			fmt.Errorf("workflow need to be approved first")))
+	}
+
+	err = CheckUserCanOperateStep(user, workflow, int(currentStep.ID))
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, err))
+	}
+	var curTaskRecord *model.WorkflowInstanceRecord
+	for _, ir := range workflow.Record.InstanceRecords {
+		if uint64(ir.TaskId) == taskIdUint {
+			curTaskRecord = ir
+		}
+	}
+	if curTaskRecord == nil {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, _err.New("task is not found in workflow")))
+	}
+
+	if req.ScheduleTime != nil && req.ScheduleTime.Before(time.Now()) {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, fmt.Errorf(
+			"request schedule time is too early")))
+	}
+
+	if curTaskRecord.IsSQLExecuted {
+		return controller.JSONBaseErrorReq(c, errors.New(errors.DataInvalid, fmt.Errorf(
+			"task has been executed")))
+	}
+
+	instance, exist, err := s.GetInstanceById(fmt.Sprintf("%v", curTaskRecord.InstanceId))
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, ErrInstanceNotExist)
+	}
+
+	if req.ScheduleTime != nil && len(instance.MaintenancePeriod) != 0 && !instance.MaintenancePeriod.IsWithinScope(*req.ScheduleTime) {
+		return controller.JSONBaseErrorReq(c, ErrWorkflowExecuteTimeIncorrect)
+	}
+
+	err = s.UpdateInstanceRecordSchedule(curTaskRecord, user.ID, req.ScheduleTime)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	return c.JSON(http.StatusOK, controller.NewBaseReq(nil))
+}
+
+// ExecuteTasksOnWorkflowV1
+// @Summary 多数据源批量上线
+// @Description execute tasks on workflow
+// @Tags workflow
+// @Id executeTasksOnWorkflowV1
+// @Security ApiKeyAuth
+// @Param workflow_id path string true "workflow id"
+// @Success 200 {object} controller.BaseRes
+// @router /v1/projects/workflows/{workflow_id}/tasks/execute [post]
+func ExecuteTasksOnWorkflowV1(c echo.Context) error {
+	workflowId := c.Param("workflow_id")
+	id, err := FormatStringToInt(workflowId)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+
+	s := model.GetStorage()
+	workflow, exist, err := s.GetWorkflowDetailById(strconv.Itoa(id))
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if !exist {
+		return controller.JSONBaseErrorReq(c, ErrWorkflowNoAccess)
+	}
+	user, err := controller.GetCurrentUser(c)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	if err := PrepareForWorkflowExecution(c, workflow, user, id); err != nil {
+		return err
+	}
+
+	needExecTaskIds, err := GetNeedExecTaskIds(s, workflow, user)
+	if err != nil {
+		return err
+	}
+
+	err = server.ExecuteWorkflow(workflow, needExecTaskIds)
+	if err != nil {
+		return controller.JSONBaseErrorReq(c, err)
+	}
+	return c.JSON(http.StatusOK, controller.NewBaseReq(nil))
 }
